+++ ops/csrc/fp8/deep_gemm/jit/compiler.py
@@ -36,9 +36,15 @@ runtime_cache = RuntimeCache()
 
 
 def hash_to_hex(s: str) -> str:
-    md5 = hashlib.md5()
-    md5.update(s.encode("utf-8"))
-    return md5.hexdigest()[0:12]
+    """Transforms input into a hexadecimal hash string"""
+    if isinstance(s, int):
+        return f"{s & 0xFFFFFFFFFFFFFFFF:016x}"
+    elif isinstance(s, str):
+        md5 = hashlib.md5()
+        md5.update(s.encode("utf-8"))
+        return md5.hexdigest()[0:12]
+    else:
+        raise TypeError(f"Unsupported type for hashing: {type(s)}")
 
 
 @functools.lru_cache(maxsize=None)
@@ -149,13 +155,41 @@ class Compiler:
         return [get_jit_include_dir()]
 
     @classmethod
-    def build(cls, name: str, code: str, runtime_cls: Type[Runtime], kwargs: Dict[str, Any] = None) -> Runtime:
+    def build(cls, name: str, runtime_cls: Type[Runtime], kwargs: Dict[str, Any] = None) -> Runtime:
         # Compiler flags
         flags = cls.flags()
-
         # Build signature
         enable_sass_opt = cls.__version__() <= (12, 8) and not int(os.getenv("DG_JIT_DISABLE_FFMA_INTERLEAVE", 0))
-        signature = f"{name}$${get_deep_gemm_version()}$${cls.signature()}$${flags}$${enable_sass_opt}$${code}"
+
+        def fast_dict_hash(kwargs: dict) -> str:
+            keys = [
+                "N",
+                "K",
+                "BLOCK_M",
+                "BLOCK_N",
+                "BLOCK_K",
+                "BLOCK_N_PADDING",
+                "SWIZZLE_D_MODE",
+                "NUM_GROUPS",
+                "NUM_STAGES",
+                "NUM_TMA_THREADS",
+                "NUM_MATH_THREADS_PER_GROUP",
+                "NUM_TMA_MULTICAST",
+                "IS_TMA_MULTICAST_ON_A",
+                "GEMM_TYPE",
+            ]
+            hashable = tuple((k, kwargs[k]) for k in keys)
+            return hash_to_hex(hash(hashable))
+
+        signature_parts = [
+            f"{name}",
+            f"{get_deep_gemm_version()}",
+            f"{cls.signature()}",
+            f"{flags}",
+            f"{enable_sass_opt}",
+            f"{fast_dict_hash(kwargs)}",
+        ]
+        signature = "$$".join(signature_parts)
         name = f"kernel.{name}.{hash_to_hex(signature)}"
         path = os.path.join(get_cache_dir(), name)
 
@@ -172,6 +206,7 @@ class Compiler:
         cubin_path = os.path.join(path, "kernel.cubin")
         tmp_cubin_path = os.path.join(make_tmp_dir(), f"nvcc.tmp.{str(uuid.uuid4())}.{hash_to_hex(cubin_path)}.cubin")
 
+        code = runtime_cls.generate(kwargs)
         start_time = time.time()
         cls.compile(name, code, tmp_cubin_path)
         end_time = time.time()
@@ -309,6 +344,6 @@ class NVRTCCompiler(Compiler):
         ), f"Failed to destroy program: {result}"
 
 
-def build(name: str, code: str, runtime_cls: Type[Runtime], kwargs: Dict[str, Any] = None) -> Runtime:
+def build(name: str, runtime_cls: Type[Runtime], kwargs: Dict[str, Any] = None) -> Runtime:
     compiler_cls = NVRTCCompiler if int(os.getenv("DG_JIT_USE_NVRTC", 0)) else NVCCCompiler
-    return compiler_cls.build(name, code, runtime_cls, kwargs)
+    return compiler_cls.build(name, runtime_cls, kwargs)
