+++ ops/csrc/fp8/deep_gemm/jit/runtime.py
@@ -16,6 +16,8 @@
 # Copyright (c) 2025 DeepSeek
 # Licensed under the MIT License - https://github.com/deepseek-ai/DeepGEMM/blob/main/LICENSE
 
+import ctypes
+import enum
 import os
 import subprocess
 import time
@@ -24,6 +26,28 @@ from typing import Any, Dict, Optional, Type
 import cuda.bindings.driver as cbd
 import paddle
 from paddle.utils.cpp_extension.cpp_extension import CUDA_HOME
+
+
+def get_num_math_warpgroups(block_m: int) -> int:
+    return 1 if block_m == 64 else 2
+
+
+def get_num_threads_per_sm(num_tma_threads: int, num_math_threads_per_group: int, block_m: int) -> int:
+    assert num_math_threads_per_group == 128, "Only support 128 threads per math group"
+    return get_num_math_warpgroups(block_m) * num_math_threads_per_group + num_tma_threads
+
+
+class GemmType(enum.Enum):
+    Normal = 0
+    GroupedContiguous = 1
+    GroupedMasked = 2
+
+    def __str__(self) -> str:
+        return {
+            0: "Normal",
+            1: "GroupedContiguous",
+            2: "GroupedMasked",
+        }[self.value]
 
 
 class Runtime:
@@ -129,3 +153,225 @@ class RuntimeCache:
             self.cache[path] = runtime
             return runtime
         return None
+
+
+def get_cache_key(kwargs, num_tma_threads, num_math_threads_per_group):
+    key_params = {
+        "NUM_TMA_MULTICAST": kwargs["NUM_TMA_MULTICAST"],
+        "NUM_SMS": kwargs["NUM_SMS"],
+        "BLOCK_M": kwargs["BLOCK_M"],
+        "SMEM_SIZE": kwargs["SMEM_SIZE"],
+        "STREAM": kwargs["STREAM"],
+        "num_tma_threads": num_tma_threads,
+        "num_math_threads_per_group": num_math_threads_per_group,
+    }
+    return hash(frozenset(key_params.items()))
+
+
+class KernelLaunchCache:
+    def __init__(self):
+        self.config_cache = {}
+        self.attr_cache = {}
+
+    @staticmethod
+    def create_attr(kwargs):
+        """Creates and caches property objects"""
+        attr_val = cbd.CUlaunchAttributeValue()
+        attr_val.clusterDim.x = kwargs["NUM_TMA_MULTICAST"]
+        attr_val.clusterDim.y = 1
+        attr_val.clusterDim.z = 1
+        attr = cbd.CUlaunchAttribute()
+        attr.id = cbd.CUlaunchAttributeID.CU_LAUNCH_ATTRIBUTE_CLUSTER_DIMENSION
+        attr.value = attr_val
+        return attr
+
+    @staticmethod
+    def create_config(kwargs, num_tma_threads, num_math_threads_per_group, attr):
+        """Creates and caches configuration objects"""
+        config = cbd.CUlaunchConfig()
+        config.numAttrs = 1
+        config.attrs = [attr]
+        config.gridDimX = kwargs["NUM_SMS"]
+        config.gridDimY = 1
+        config.gridDimZ = 1
+        config.blockDimX = get_num_threads_per_sm(num_tma_threads, num_math_threads_per_group, kwargs["BLOCK_M"])
+        config.blockDimY = 1
+        config.blockDimZ = 1
+        config.sharedMemBytes = kwargs["SMEM_SIZE"]
+        config.hStream = kwargs["STREAM"]
+        return config
+
+    def get_launch_config(self, kwargs, num_tma_threads, num_math_threads_per_group):
+        """Retrieves cached config or creates new instance"""
+        cache_key = get_cache_key(kwargs, num_tma_threads, num_math_threads_per_group)
+
+        if cache_key not in self.config_cache:
+            # 先检查属性缓存
+            attr_key = (kwargs["NUM_TMA_MULTICAST"],)
+            if attr_key not in self.attr_cache:
+                self.attr_cache[attr_key] = self.create_attr(kwargs)
+
+            # 创建新配置
+            config = self.create_config(kwargs, num_tma_threads, num_math_threads_per_group, self.attr_cache[attr_key])
+            self.config_cache[cache_key] = config
+
+        return self.config_cache[cache_key]
+
+
+launch_cache = KernelLaunchCache()
+
+
+class FP8GemmRuntime(Runtime):
+    def __init__(self, path: str) -> None:
+        super().__init__(path)
+
+    @staticmethod
+    def generate(kwargs: Dict[str, Any]) -> str:
+        code = f"""
+#ifdef __CUDACC_RTC__
+#include <deep_gemm/nvrtc_std.cuh>
+#else
+#include <cuda.h>
+#include <string>
+#endif
+
+#include <cuda_bf16.h>
+#include <cuda_fp8.h>
+
+#include <deep_gemm/fp8_gemm.cuh>
+
+using namespace deep_gemm;
+
+static void __instantiate_kernel() {{
+    auto ptr = reinterpret_cast<void*>(&fp8_gemm_kernel<
+        {kwargs['N']},
+        {kwargs['K']},
+        {kwargs['BLOCK_M']},
+        {kwargs['BLOCK_N']},
+        {kwargs['BLOCK_K']},
+        {kwargs['BLOCK_N_PADDING']},
+        {kwargs['SWIZZLE_D_MODE']},
+        {kwargs['NUM_GROUPS']},
+        {kwargs['NUM_STAGES']},
+        {kwargs['NUM_TMA_THREADS']},
+        {kwargs['NUM_MATH_THREADS_PER_GROUP']},
+        {kwargs['NUM_TMA_MULTICAST']},
+        {'true' if kwargs['IS_TMA_MULTICAST_ON_A'] else 'false'},
+        GemmType::{kwargs['GEMM_TYPE']}
+      >);
+}};
+"""
+        if int(os.getenv("DG_JIT_DEBUG", 0)):
+            print(f"Generated FP8 GEMM code:\n{code}")
+        return code
+
+    # noinspection PyMethodOverriding
+    @staticmethod
+    def launch(kernel: cbd.CUkernel, kwargs: Dict[str, Any]) -> cbd.CUresult:
+        num_tma_threads = 128
+        num_math_threads_per_group = 128
+
+        result = cbd.cuKernelSetAttribute(
+            cbd.CUfunction_attribute.CU_FUNC_ATTRIBUTE_MAX_DYNAMIC_SHARED_SIZE_BYTES,
+            kwargs["SMEM_SIZE"],
+            kernel,
+            cbd.CUdevice(kwargs["DEVICE_INDEX"]),
+        )[0]
+        assert result == cbd.CUresult.CUDA_SUCCESS, f"Failed to set max dynamic shared memory size: {result}"
+        config = launch_cache.get_launch_config(kwargs, num_tma_threads, num_math_threads_per_group)
+
+        arg_values = (
+            kwargs["SCALES_B"].data_ptr(),
+            kwargs["GROUPED_LAYOUT"].data_ptr(),
+            kwargs["M"],
+            kwargs["TENSOR_MAP_A"],
+            kwargs["TENSOR_MAP_B"],
+            kwargs["TENSOR_MAP_SCALES_A"],
+            kwargs["TENSOR_MAP_D"],
+        )
+        arg_types = (
+            ctypes.c_void_p,
+            ctypes.c_void_p,
+            ctypes.c_uint32,
+            None,
+            None,
+            None,
+            None,
+        )
+        ret = cbd.cuLaunchKernelEx(config, kernel, (arg_values, arg_types), 0)
+        return ret
+
+
+class FP8WGradGemmRuntime(Runtime):
+    def __init__(self, path: str) -> None:
+        super().__init__(path)
+
+    @staticmethod
+    def generate(kwargs: Dict[str, Any]) -> str:
+        code = f"""
+#ifdef __CUDACC_RTC__
+#include <deep_gemm/nvrtc_std.cuh>
+#else
+#include <cuda.h>
+#include <string>
+#endif
+
+#include <cuda_bf16.h>
+#include <cuda_fp8.h>
+
+#include <deep_gemm/fp8_wgrad_gemm.cuh>
+
+using namespace deep_gemm;
+
+static void __instantiate_kernel() {{
+    auto ptr = reinterpret_cast<void*>(&fp8_wgrad_gemm_kernel<
+        {kwargs['M']},
+        {kwargs['N']},
+        {kwargs['BLOCK_M']},
+        {kwargs['BLOCK_N']},
+        {kwargs['BLOCK_K']},
+        {kwargs['NUM_STAGES']},
+        {kwargs['NUM_LAST_STAGES']},
+        {kwargs['NUM_TMA_THREADS']},
+        {kwargs['NUM_MATH_THREADS_PER_GROUP']},
+        {kwargs['NUM_TMA_MULTICAST']},
+        {'true' if kwargs['IS_TMA_MULTICAST_ON_A'] else 'false'}
+      >);
+}};
+"""
+        if int(os.getenv("DG_JIT_DEBUG", 0)):
+            print(f"Generated FP8 WGrad GEMM code:\n{code}")
+        return code
+
+    # noinspection PyMethodOverriding
+    @staticmethod
+    def launch(kernel: cbd.CUkernel, kwargs: Dict[str, Any]) -> cbd.CUresult:
+        num_tma_threads = 128
+        num_math_threads_per_group = 128
+
+        result = cbd.cuKernelSetAttribute(
+            cbd.CUfunction_attribute.CU_FUNC_ATTRIBUTE_MAX_DYNAMIC_SHARED_SIZE_BYTES,
+            kwargs["SMEM_SIZE"],
+            kernel,
+            cbd.CUdevice(kwargs["DEVICE_INDEX"]),
+        )[0]
+        assert result == cbd.CUresult.CUDA_SUCCESS, f"Failed to set max dynamic shared memory size: {result}"
+        config = launch_cache.get_launch_config(kwargs, num_tma_threads, num_math_threads_per_group)
+
+        arg_values = (
+            kwargs["K"],
+            kwargs["TENSOR_MAP_A"],
+            kwargs["TENSOR_MAP_B"],
+            kwargs["TENSOR_MAP_SCALES_A"],
+            kwargs["TENSOR_MAP_SCALES_B"],
+            kwargs["TENSOR_MAP_D"],
+        )
+        arg_types = (
+            ctypes.c_uint32,
+            None,
+            None,
+            None,
+            None,
+            None,
+        )
+        return cbd.cuLaunchKernelEx(config, kernel, (arg_values, arg_types), 0)
