+++ ops/csrc/fp8/deep_gemm/jit_kernels/runtime.py
@@ -16,30 +16,13 @@
 # Copyright (c) 2025 DeepSeek
 # Licensed under the MIT License - https://github.com/deepseek-ai/DeepGEMM/blob/main/LICENSE
 
-import ctypes
-import enum
-import os
 from typing import Any, Dict, Tuple
 
 import cuda.bindings.driver as cbd
 import paddle
 
-from ..jit.runtime import Runtime
+from ..jit.runtime import GemmType
 from .utils import get_tma_aligned_size
-
-
-class GemmType(enum.Enum):
-    Normal = 0
-    GroupedContiguous = 1
-    GroupedMasked = 2
-
-    def __str__(self) -> str:
-        return {
-            0: "Normal",
-            1: "GroupedContiguous",
-            2: "GroupedMasked",
-        }[self.value]
-
 
 # TODO Support dtype in Paddle
 tmap_type_map: Dict[Any, str] = {
@@ -68,15 +51,6 @@ swizzle_type_map = {
 }
 
 
-def get_num_math_warpgroups(block_m: int) -> int:
-    return 1 if block_m == 64 else 2
-
-
-def get_num_threads_per_sm(num_tma_threads: int, num_math_threads_per_group: int, block_m: int) -> int:
-    assert num_math_threads_per_group == 128, "Only support 128 threads per math group"
-    return get_num_math_warpgroups(block_m) * num_math_threads_per_group + num_tma_threads
-
-
 def make_2d_tma_copy_desc(
     t: paddle.Tensor,
     gmem_dims: Tuple[cbd.cuuint64_t, cbd.cuuint64_t],
@@ -98,7 +72,6 @@ def make_2d_tma_copy_desc(
         cbd.CUtensorMapL2promotion.CU_TENSOR_MAP_L2_PROMOTION_L2_256B,
         cbd.CUtensorMapFloatOOBfill.CU_TENSOR_MAP_FLOAT_OOB_FILL_NONE,
     )
-
     if res != cbd.CUresult.CUDA_SUCCESS:
         raise Exception(f"Failed to encode tensor map: {res}")
     return tensor_map
@@ -188,196 +161,3 @@ def make_2d_tma_scales_desc(
         1,
         cbd.CUtensorMapSwizzle.CU_TENSOR_MAP_SWIZZLE_NONE,
     )
-
-
-class FP8GemmRuntime(Runtime):
-    def __init__(self, path: str) -> None:
-        super().__init__(path)
-
-    @staticmethod
-    def generate(kwargs: Dict[str, Any]) -> str:
-        code = f"""
-#ifdef __CUDACC_RTC__
-#include <deep_gemm/nvrtc_std.cuh>
-#else
-#include <cuda.h>
-#include <string>
-#endif
-
-#include <cuda_bf16.h>
-#include <cuda_fp8.h>
-
-#include <deep_gemm/fp8_gemm.cuh>
-
-using namespace deep_gemm;
-
-static void __instantiate_kernel() {{
-    auto ptr = reinterpret_cast<void*>(&fp8_gemm_kernel<
-        {kwargs['N']},
-        {kwargs['K']},
-        {kwargs['BLOCK_M']},
-        {kwargs['BLOCK_N']},
-        {kwargs['BLOCK_K']},
-        {kwargs['BLOCK_N_PADDING']},
-        {kwargs['SWIZZLE_D_MODE']},
-        {kwargs['NUM_GROUPS']},
-        {kwargs['NUM_STAGES']},
-        {kwargs['NUM_TMA_THREADS']},
-        {kwargs['NUM_MATH_THREADS_PER_GROUP']},
-        {kwargs['NUM_TMA_MULTICAST']},
-        {'true' if kwargs['IS_TMA_MULTICAST_ON_A'] else 'false'},
-        GemmType::{kwargs['GEMM_TYPE']}
-      >);
-}};
-"""
-        if int(os.getenv("DG_JIT_DEBUG", 0)):
-            print(f"Generated FP8 GEMM code:\n{code}")
-        return code
-
-    # noinspection PyMethodOverriding
-    @staticmethod
-    def launch(kernel: cbd.CUkernel, kwargs: Dict[str, Any]) -> cbd.CUresult:
-        num_tma_threads = 128
-        num_math_threads_per_group = 128
-
-        result = cbd.cuKernelSetAttribute(
-            cbd.CUfunction_attribute.CU_FUNC_ATTRIBUTE_MAX_DYNAMIC_SHARED_SIZE_BYTES,
-            kwargs["SMEM_SIZE"],
-            kernel,
-            cbd.CUdevice(kwargs["DEVICE_INDEX"]),
-        )[0]
-        assert result == cbd.CUresult.CUDA_SUCCESS, f"Failed to set max dynamic shared memory size: {result}"
-
-        attr_val = cbd.CUlaunchAttributeValue()
-        attr_val.clusterDim.x = kwargs["NUM_TMA_MULTICAST"]
-        attr_val.clusterDim.y = 1
-        attr_val.clusterDim.z = 1
-        attr = cbd.CUlaunchAttribute()
-        attr.id = cbd.CUlaunchAttributeID.CU_LAUNCH_ATTRIBUTE_CLUSTER_DIMENSION
-        attr.value = attr_val
-
-        config = cbd.CUlaunchConfig()
-        config.numAttrs = 1
-        config.attrs = [attr]
-        config.gridDimX = kwargs["NUM_SMS"]
-        config.gridDimY = 1
-        config.gridDimZ = 1
-        config.blockDimX = get_num_threads_per_sm(num_tma_threads, num_math_threads_per_group, kwargs["BLOCK_M"])
-        config.blockDimY = 1
-        config.blockDimZ = 1
-        config.sharedMemBytes = kwargs["SMEM_SIZE"]
-        config.hStream = kwargs["STREAM"]
-
-        arg_values = (
-            kwargs["SCALES_B"].data_ptr(),
-            kwargs["GROUPED_LAYOUT"].data_ptr(),
-            kwargs["M"],
-            kwargs["TENSOR_MAP_A"],
-            kwargs["TENSOR_MAP_B"],
-            kwargs["TENSOR_MAP_SCALES_A"],
-            kwargs["TENSOR_MAP_D"],
-        )
-        arg_types = (
-            ctypes.c_void_p,
-            ctypes.c_void_p,
-            ctypes.c_uint32,
-            None,
-            None,
-            None,
-            None,
-        )
-        return cbd.cuLaunchKernelEx(config, kernel, (arg_values, arg_types), 0)
-
-
-class FP8WGradGemmRuntime(Runtime):
-    def __init__(self, path: str) -> None:
-        super().__init__(path)
-
-    @staticmethod
-    def generate(kwargs: Dict[str, Any]) -> str:
-        code = f"""
-#ifdef __CUDACC_RTC__
-#include <deep_gemm/nvrtc_std.cuh>
-#else
-#include <cuda.h>
-#include <string>
-#endif
-
-#include <cuda_bf16.h>
-#include <cuda_fp8.h>
-
-#include <deep_gemm/fp8_wgrad_gemm.cuh>
-
-using namespace deep_gemm;
-
-static void __instantiate_kernel() {{
-    auto ptr = reinterpret_cast<void*>(&fp8_wgrad_gemm_kernel<
-        {kwargs['M']},
-        {kwargs['N']},
-        {kwargs['BLOCK_M']},
-        {kwargs['BLOCK_N']},
-        {kwargs['BLOCK_K']},
-        {kwargs['NUM_STAGES']},
-        {kwargs['NUM_LAST_STAGES']},
-        {kwargs['NUM_TMA_THREADS']},
-        {kwargs['NUM_MATH_THREADS_PER_GROUP']},
-        {kwargs['NUM_TMA_MULTICAST']},
-        {'true' if kwargs['IS_TMA_MULTICAST_ON_A'] else 'false'}
-      >);
-}};
-"""
-        if int(os.getenv("DG_JIT_DEBUG", 0)):
-            print(f"Generated FP8 WGrad GEMM code:\n{code}")
-        return code
-
-    # noinspection PyMethodOverriding
-    @staticmethod
-    def launch(kernel: cbd.CUkernel, kwargs: Dict[str, Any]) -> cbd.CUresult:
-        num_tma_threads = 128
-        num_math_threads_per_group = 128
-
-        result = cbd.cuKernelSetAttribute(
-            cbd.CUfunction_attribute.CU_FUNC_ATTRIBUTE_MAX_DYNAMIC_SHARED_SIZE_BYTES,
-            kwargs["SMEM_SIZE"],
-            kernel,
-            cbd.CUdevice(kwargs["DEVICE_INDEX"]),
-        )[0]
-        assert result == cbd.CUresult.CUDA_SUCCESS, f"Failed to set max dynamic shared memory size: {result}"
-
-        attr_val = cbd.CUlaunchAttributeValue()
-        attr_val.clusterDim.x = kwargs["NUM_TMA_MULTICAST"]
-        attr_val.clusterDim.y = 1
-        attr_val.clusterDim.z = 1
-        attr = cbd.CUlaunchAttribute()
-        attr.id = cbd.CUlaunchAttributeID.CU_LAUNCH_ATTRIBUTE_CLUSTER_DIMENSION
-        attr.value = attr_val
-
-        config = cbd.CUlaunchConfig()
-        config.numAttrs = 1
-        config.attrs = [attr]
-        config.gridDimX = kwargs["NUM_SMS"]
-        config.gridDimY = 1
-        config.gridDimZ = 1
-        config.blockDimX = get_num_threads_per_sm(num_tma_threads, num_math_threads_per_group, kwargs["BLOCK_M"])
-        config.blockDimY = 1
-        config.blockDimZ = 1
-        config.sharedMemBytes = kwargs["SMEM_SIZE"]
-        config.hStream = kwargs["STREAM"]
-
-        arg_values = (
-            kwargs["K"],
-            kwargs["TENSOR_MAP_A"],
-            kwargs["TENSOR_MAP_B"],
-            kwargs["TENSOR_MAP_SCALES_A"],
-            kwargs["TENSOR_MAP_SCALES_B"],
-            kwargs["TENSOR_MAP_D"],
-        )
-        arg_types = (
-            ctypes.c_uint32,
-            None,
-            None,
-            None,
-            None,
-            None,
-        )
-        return cbd.cuLaunchKernelEx(config, kernel, (arg_values, arg_types), 0)
